#7 [REFACTOR] API 함수 분할 및 주석 추가
